Extracted a FakeStorage class ...and use it for both fake session and local storage
