+++ src/test/EmailAuthentication.integration.test.tsx
@@ -2,7 +2,7 @@ import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import nock from 'nock';
 
-import { fakeAccessToken } from './utils/fake-local-storage';
+import { fakeAccessToken } from './utils/storage';
 import { Mockchain } from './utils/mockchain';
 import { BASE_URL } from '_src/shared/constants';
 import { renderApp } from '_src/test/utils/react-rendering';
