+++ src/test/NavBar.integration.test.tsx
@@ -1,7 +1,7 @@
 import { screen, within } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import { simulateMnemonicUser } from './utils/fake-local-storage';
+import { simulateMnemonicUser } from './utils/storage';
 import { Mockchain } from '_src/test/utils/mockchain';
 import { renderApp } from '_src/test/utils/react-rendering';
