+++ src/test/Permissions.integration.test.tsx
@@ -4,7 +4,7 @@ import userEvent from '@testing-library/user-event';
 import {
     simulateConnectedApps,
     simulateMnemonicUser,
-} from '_src/test/utils/fake-local-storage';
+} from '_src/test/utils/storage';
 import { Mockchain } from '_src/test/utils/mockchain';
 import { renderApp } from '_src/test/utils/react-rendering';
