+++ src/test/Security.integration.test.tsx
@@ -11,7 +11,7 @@ import {
     recoveryPhrase,
     simulateMnemonicUser,
     simulateEmailUser,
-} from '_src/test/utils/fake-local-storage';
+} from '_src/test/utils/storage';
 import { Mockchain } from '_src/test/utils/mockchain';
 import { renderApp } from '_src/test/utils/react-rendering';
