+++ src/test/utils/fake-browser.ts
@@ -1,11 +1,12 @@
 // This module provides fake implementations of the webextension-polyfill Browser object
 // NOTE: this is incomplete and provides just enough implementation for tests to pass
 
-class FakeStorage {
-
+export class FakeStorage {
     records: Record<string, unknown> = {};
 
-    get(dkeys?: null | string | string[] | Record<string, unknown>): Promise<Record<string, unknown>> {
+    get(
+        dkeys?: null | string | string[] | Record<string, unknown>
+    ): Promise<Record<string, unknown>> {
         return new Promise<Record<string, unknown>>((resolve, reject) => {
             const returnVal: Record<string, unknown> = {};
             if (typeof dkeys === 'string') {
@@ -15,9 +16,7 @@ class FakeStorage {
         });
     }
 
-    async set(
-        items: Record<string, unknown>
-    ): Promise<void> {
+    async set(items: Record<string, unknown>): Promise<void> {
         for (const property in items) {
             this.records[property] = items[property];
         }
@@ -27,12 +26,12 @@ class FakeStorage {
     }
 
     clear() {
-        this.records = {}
+        this.records = {};
     }
-
 }
 
 const fakeLocalStorage = new FakeStorage();
+export const fakeSessionStorage = new FakeStorage();
 
 function fakeEvent() {
     return {
@@ -42,6 +41,7 @@ function fakeEvent() {
         hasListeners: () => false,
     };
 }
+
 function fakePort() {
     return {
         name: 'whatever',
@@ -51,6 +51,7 @@ function fakePort() {
         postMessage: jest.fn(),
     };
 }
+
 export const fakeBrowser = {
     runtime: {
         id: 'chrome-runtime-id',
@@ -62,6 +63,6 @@ export const fakeBrowser = {
     },
 };
 
-export const clearLocalStorage = () => {
+export const clearFakeStorages = () => {
     fakeLocalStorage.clear();
 };
